clean up after not using the lamp anymore

+++ scripts/setup.sh
@@ -1,3 +1,12 @@
+if [ ! -f ./credentials.txt ]
+then
+  echo "Cannot start without credentials. See readme for more details"
+  exit
+fi
+
+# install dependencies
+pip3 install -r requirements.txt
+
 # create, enable and start service
 sudo cp ./scripts/alarm.service /lib/systemd/system/alarm.service
 sudo systemctl daemon-reload
